lotta stuff
largely this is the half way mark for making the table and doing rando checkm stuff (it would have been wise to do them seperately but i was waiting and figured i should use the time productively). so now there is a directory of flye stuff and some script changes on some documents. still need to make the table, find out just what in the hell the outputs from checkm mean and notebooking.